move upgrades.js to ./src
